backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
motor>=3.7,<4
pymongo>=4.9,<5
pydantic==2.7.1
python-dotenv==1.0.1
dnspython==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
numpy>=1.23
python-multipart>=0.0.9
